testing mobile app

+++ android/app/src/main/java/com/spending/intelligence/presentation/budgets/BudgetsScreen.kt
@@ -89,7 +89,7 @@ private fun GoalCard(goal: Goal) {
             }
             Spacer(Modifier.height(6.dp))
             LinearProgressIndicator(
-                progress = { (goal.progressPercentage / 100f).coerceIn(0f, 1f) },
+                progress = { (goal.progressPercentage.toFloat() / 100f).coerceIn(0f, 1f) },
                 modifier = Modifier.fillMaxWidth().height(6.dp),
                 color = predictionColor
             )
